Improve accessibilty for anchor link navigation
